FFmpeg subs update
